pyproject.toml
[project]
name = "ComfyUI_M3D_photo_effects"
description = "A set of photo effects for ComfyUI, like Bleach Bypass or RGB curve. More to come!"
version = "1.1.0"
license = { file = "LICENSE" }
dependencies = ["Pillow", "numpy"]

[project.urls]
Repository = "https://github.com/metal3d/ComfyUI_M3D_photo_effects"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "metal3d"
DisplayName = "M3D photo effects"
Icon = ""
